AppsListing refactored to AppsNav, uses QueryWrapper
