+++ src/tapis-ui/components/QueryWrapper/index.ts
@@ -1,0 +1,3 @@
+import { default as QueryWrapper } from './QueryWrapper';
+
+export default QueryWrapper;
