--- src/tapis-ui/components/apps/AppsListing/index.ts
@@ -1,2 +1,0 @@
-import AppsListing from './AppsListing';
-export default AppsListing;
