+++ src/tapis-ui/components/apps/AppsNav/AppsNav.test.tsx
@@ -1,15 +1,14 @@
 import React from 'react';
-import configureStore from 'redux-mock-store';
 import '@testing-library/jest-dom/extend-expect';
 import renderComponent from 'utils/testing';
-import { AppsListing } from 'tapis-ui/components/apps';
+import { AppsNav } from 'tapis-ui/components/apps';
 import { useList } from 'tapis-hooks/apps';
 import { tapisApp } from 'fixtures/apps.fixtures';
 
 jest.mock('tapis-hooks/apps');
 
-describe('Apps', () => {
-  it('renders AppsListing component', () => {
+describe('AppsNav', () => {
+  it('renders AppsNav component', () => {
     (useList as jest.Mock).mockReturnValue({
       data: {
         result: [ tapisApp ],
@@ -18,7 +17,7 @@ describe('Apps', () => {
       error: null
     })
 
-    const { getAllByText } = renderComponent(<AppsListing />);
+    const { getAllByText } = renderComponent(<AppsNav />);
     expect(getAllByText(/SleepSeconds/).length).toEqual(1);
   });
 });
