+++ src/tapis-ui/components/apps/AppsNav/index.ts
@@ -1,0 +1,2 @@
+import AppsNav from './AppsNav';
+export default AppsNav;
