+++ src/tapis-ui/components/apps/index.ts
@@ -1,1 +1,1 @@
-export { default as AppsListing } from './AppsListing';
+export { default as AppsNav } from './AppsNav';
